refactor: command API 컨트롤러를 기능별로 분리 - SellerChannelCommandController.connectSellerChannel 셀러 채널 연결 엔드포인트를 전용 컨트롤러로 이동함 - OrderSyncCommandController.syncByChannelKey 주문 동기화·가져오기·미리보기 경로를 주문 동기화 컨트롤러로 분리함 - FulfillmentCommandController.createBulkFulfillment fulfillment 단건·일괄 API 책임을 별도 컨트롤러로 고립함 - InvoiceCommandController.createManualOrderInvoice 송장 발급 관련 엔드포인트를 전용 컨트롤러로 재구성함
